+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [0.9.22] - 2019-02-27
 ### Added
 - Added a setting for including hideouts and towns for gain per area (off by default)
+- Added pricing of 6-sockets (now converts 6-sockets to jewellers automatically)
 
 ### Fixed
 - Fixed a bug where the inventory was not fetched correctly directly when logging in